fix(Dropdown): 显示选中选项时，预留选中图标空间
