+++ components/dropdown/index.ts
@@ -6,7 +6,7 @@ import type { SFCWithInstall } from '../_util/interface';
 type DropdownType = SFCWithInstall<typeof Dropdown>;
 export const FDropdown = withInstall<DropdownType>(Dropdown as DropdownType);
 
-export { dropdownProps } from './dropdown';
-export type { DropdownProps } from './dropdown';
+export { dropdownProps } from './props';
+export type { DropdownProps } from './props';
 
 export default FDropdown;
